% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsview-tsplot.R
\name{tsplot}
\alias{tsplot}
\title{Plotting Time Series Objects}
\usage{
tsplot(x = ts(matrix(rnorm(150), 30, 5), start = c(1961, 1), frequency = 4),
  plot.type = "multiple")
}
\arguments{
\item{x}{time series object, usually inheriting from class "\code{ts}" or "\code{regts}".}

\item{plot.type}{for multivariate time series. \code{multiple} displays each series separately (with a common time axis), \code{single} displays all series in the same plot.}
}
\description{
Plotting method for objects inheriting from classes "\code{ts}" and "\code{regts}".
}
\examples{
x = ts(matrix(rnorm(150), 30, 5), start = c(1961, 1), frequency = 4) # 5 time series

tsplot(x, "single")
tsplot(x, "multiple")

}
\seealso{
\code{\link{tsview}, \link{ts}, \link{regts}, \link{grepl}}
}

