% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsview.R
\name{tsview}
\alias{tsview}
\title{Plotting Time Series Objects (in Browser)}
\usage{
tsview(x = ts(matrix(rnorm(150), 30, 5), start = c(1961, 1), frequency = 4),
  plot.type = "multiple", gui = FALSE)
}
\arguments{
\item{x}{time series object, usually inheriting from class "\code{ts}".}

\item{plot.type}{for multivariate time series. \code{multiple} displays each series separately (with a common time axis), \code{single} displays all series in the same plot.}

\item{gui}{if gui = FALSE time series are shown in a regular plot. Else, if gui = TRUE, plots are shown in a gui (c.q. your default web browser).}
}
\description{
Plotting method for objects inheriting from class "\code{ts}".
}
\examples{
x = ts(matrix(rnorm(150), 30, 5), start = c(1961, 1), frequency = 4) # 5 time series

tsview(x, "single")
tsview(x, "multiple")

\dontrun{
tsview(x, gui = TRUE)
}
}
\seealso{
\code{\link{ts}, \link{grepl}}
}

